Windows fixes, 32 bit compatibility

+++ test/parametric.jl
@@ -35,5 +35,5 @@ nontype2 = ParametricTypes.NonTypeParam{UInt32, UInt32(2)}()
 nontype3 = ParametricTypes.NonTypeParam{Int32, Int32(1)}(3)
 @test ParametricTypes.get_nontype(nontype3) == 3
 
-nontype4 = ParametricTypes.NonTypeParam{Int64, 64}()
-@test ParametricTypes.get_nontype(nontype4) == 64
+nontype4 = ParametricTypes.NonTypeParam{Int64, Int64(64)}()
+@test ParametricTypes.get_nontype(nontype4) == Int64(64)
